feat(rating): set cursor on stars of rating component
The cursor used on the stars of the rating component is a text selection cursor, which doesn't make it clear that they're clickable. The cursor stays the same if the rating component is readonly.
This commit sets a 'pointer' cursor on the stars, and a 'not-allowed' cursor when it's read-only.