demo: change the demo theme to the default one
This changes the demo theme from the dark one to the default one, and sets the background to the same color as the bootstrap card header, to make it look more integrated with bootstrap.

Also moved the vendor CSS to the vendor bundle. Not really important for our demo but I like to keep the things clean.

Fixes #414