feat(datepicker): keep day/month when adding/removing month/year in ngb-calendar
BREAKING CHANGES: ngb-calendar getNext/getPrev specifications changed

`calendar.getNext()`/`calendar.getPrev()` with the period parameter set
as `'m'`/`'y'` no longer automatically returns the first day of the
month / the first day of the year. The day in the month and the month in
the year of the input date are now kept as much as possible.
In case the destination month is is not long enough to contain a day with
the same number, the last day of the month is used.

For example:

`calendar.getNext(new NgbDate(2016, 7, 22), 'm'))`
 previously returned: `new NgbDate(2016, 8, 1))`
  and now it returns: `new NgbDate(2016, 8, 22))`

`calendar.getNext(new NgbDate(2016, 7, 22), 'y'))`
 previously returned: `new NgbDate(2017, 1, 1))`
  and now it returns: `new NgbDate(2017, 7, 22))`

`calendar.getNext(new NgbDate(2016, 1, 30), 'm'))`
 previously returned: `new NgbDate(2016, 2, 1))`
  and now it returns: `new NgbDate(2016, 2, 29))`
 (because 2016-02-30 is not a valid date)