fix(933): accept `document:click` event
BREAKING CHANGES: trigger syntax has been modified to comply with Angular's. Triggers can now be declared this way : `openTriggers/closeTriggers`.
Triggers are represented with the following syntax `(target:)event`
The target is optional and can be either `document`, `body` or`window`.

Improve perf (close events are registered dynamically, when opening a tooltip or a popover)