feat(datepicker): use the angular locale API
the datepicker now honors the Angular LOCALE_ID and uses the registered
locale data to internationalize the months and days names. Unless you
really want custom names, providing a custom NgbDatepickerI18n service
isn't necessary anymore to internationalize the datepicker.

fix #2065

BREAKING CHANGE: if your application provides a LOCALE_ID other than
the default en-US, registers the locale data for this locale, and
doesn't use a custom NgbDatepickerI18n, then the days and months
of the datepicker won't be displayed in English anymore, but in the
langguage of the provided locale.

Closes #2066