demo: change the strategy to generate random images in carousel demo
The previous strategy doesn't work in StackBlitz because the https requests redirect to http.
And asking random images to picsum.photos causes the same image to be used, probably because of of a too agressive cache. We thus generate random image paths to make sure that the 3 images are actually different.