fix(datepicker): empty month select box when displayMonths > 1
The list of months and years was generated from the focused date rather
than the first displayed date, which was inconsistent with the value used
in the datepicker navigation component, which uses the first displayed date.
This inconsistency was not visible when displayMonths = 1 but it sometimes
caused the following issue when several months were displayed:
the month select box was empty as it did not contain the currently displayed
month.

Closes #2377