fix(datepicker): ExpressionChangedAfterItHasBeenCheckedError
Reuses data structures in the datepicker data model instead of building
new objects each time the user navigates to a different month.
As a consequence, DOM elements are no longer destroyed when changing month.

This fixes the following issue which occurred when BrowserAnimationsModule was
used: when using the keyboard to navigate from one month to another, the
focused element was destroyed during the change detection of the keyboard
event, which synchronously triggered the focusout event in Chrome, leading
to a change in the data model (and that change is still part of the data
model only when BrowserAnimationsModule is used, because that module changes
the behavior of Angular to not destroy views and their data model as quickly
as when this module is not loaded), and this was causing the
ExpressionChangedAfterItHasBeenCheckedError error.

Closes #2408