fix(datepicker): access view child from ngAfterViewInit
datepicker accessed a view child from the ngAfterContentInit hook instead of accessing it from the ngAfterViewInit hook.
This doesn't work anymore with Ivy. That is a regression in Ivy, but is also an issue with ng-bootstrap which doesn't respect the standard life-cycle.
Besides, the component didn't implement the AfterContentInit interface.
It now implements the AfterViewInit interface.

fix #3150