feat(carousel): stay paused after calling pause() until cycle() is called
After setting the carousel on pause through its pause() method and clicking
on navigation arrows, the carousel no longer automatically restarts.
Calling the cycle() method is needed to undo the effect of the pause()
method.
When pauseOnHover is true, hovering on the carousel no longer uses the
pause() and cycle() methods. The hover status is stored separately, so that
calling pause() and removing the mouse from the carousel no longer restarts
it automatically.

The slide event now has a "source" property allowing to know what triggered
the event: "timer", "arrowLeft", "arrowRight" or "indicator".
The "paused" boolean status is now also included in the slide event.

Closes #3188