docs(table): fixed search for complete example
Previously if one would search for "China", no examples would show up.
Because the `matches`-function compares the lowercased name of the
country to the search term, exact matches are not getting noticed.
This commit fixes this by lowercasing the search term when it is
compared to the lowercased country name.