fix(demo): direct link to specific carousel sample not working (#3272)
The carousel samples in the demo site contain images that are loaded
asynchronously. But as their size is not reserved beforehand, they shift
the remaining content when loaded. As a consequence, the scrolling position
that was set correctly to target a specific sample before the images are
loaded is no longer correct after the images are loaded.

This commit reserves the space for the images by using the trick documented
in the following link to also keep the responsive behavior:
https://itnext.io/how-to-stop-content-jumping-when-images-load-7c915e47f576