fix(modal): cmpt based content was not scrollable
In `5.0.0` we introduced the bootstrap scrollable content feature.
This behaviour appears to be broken with modal that have content generated
from a component. Overall styling is wrong, resulting in a broken modal
not scrollable.

The fix is adding some extra styling added via a CSS classname
`component-host-scrollable`, in order to force the component
host to fit inside the parent `.model-content` flex display.

Fixes #3281