chore(progressbar): allow to specify progressbar text type
As advised on the Bootstrap documentation page when using some
background variants that is advisable to change the text color
accordingly, for example for readability purposes. The `textType` input
property allows to achieve this, adding support for binding text types
to `text-*` utility class on the progress bar.
Common scenario: progress bar with 'warning' type renders a white text on
the yellow background (assuming default BS). This change allows users to opt to
'dark' text type to improve the readability.

ref:
https://getbootstrap.com/docs/4.3/utilities/colors/#background-color