test(ie): local configuration for InternetExplorer
The project comes with confgiguration for local IE build supporting ES5
and for SouceLabs remote configuration for IE browsers, but it has no
relevant entry to allow tests on local setup. This are minimal changes
to allow local tests with IE on Windows box.
Example syntax:
```bash
yarn test --browsers IENoExtensions --configuration ie
```
There is no configuration for size of the launched IE instance, so in
order to have tests depending on the viewport size - like overlays
positioning on top/left/right/bottom, the IE instance has been
configured to run in the kiosk mode ('-k').