test: fix flaky e2e keyboard tests
Some of the components use `requestAnimationFrame()` as a part of keyboard handling.
This might make some tests flaky when doing for example:

```
await sendKey(Keys.ESC);
expect(await page.getSomeElement());
```

Expectation in this case might fail or it might not.
So this commit forces tick in the the `sendKeys()` systematically after sending the key.