demo: remove 'tabset' section and code completely
Since the documentation is now versioned, there is no need to keep it around.
We had a 'special' `NgbModule` for the demo site that included tabset source code, so we had to actually maintain the tabset code that was removed long ago from the library, but was necessary to have on the demo site.