Allow [ngbNavLink] to be used on <button>
As an alternative to the solution proposed in #4206, this is a minimal change that allows the [ngbNavLink] to be used on <button> elements as well as links. Using <button> will make the ng-bootstrap markup resemble Bootstrap 5.2's more closely, and will resolve the errors reported in axe DevTools and other automated accessibility tests specifying that role="tab" is not appropriate on <a> tags.

I have not removed the a[ngbNavLink] selector in favor of backwards compatibility, but the docs and demos should be updated to reflect that <a> ngbNavLinks will cause accessibility testing errors when used in conjunction with the [roles] directive.