demo: use `provideHttpClient` in stackblitz demos (#4522)
This change provides an HttpClient to all the stackblitz demos to ensure
they function like they do on the demo site. Currently the HttpClient is
the only provider needed, but the stackblitz template
`misc/stackblitzes-templates/main.ts.ejs` should be kept in sync with
`demo/src/main.ts` to ensure the individual examples have all their
dependencies provided.

The needed providers can also be provided to the demo modules themselves
or as context to the template renderer, but that may be more effort than
its worth and while providing the dependencies to all the examples is
overkill it will ensure they all work correctly.