fix(tooltip): align closing behavior with Bootstrap
If several triggers are set, ex. 'hover focus', make sure that all triggers are taken into account for closure and not only the first one.
Ex: mouseout on the triggering element, should not close the tooltip while the element is focused; and focus leaving the element should not close the tooltip while the mouse is still hovering it.

Fixes #3889